app/requirements.txt
Flask==2.0.2
Flask-MySQL==1.5.2
Werkzeug==2.0.2
python-dotenv
cryptography